requirements.txt
flask==2.3.3
flask-socketio==5.1.1
python-engineio==4.2.1
python-socketio==5.4.0
firebase-admin==6.4.0
eventlet==0.33.3
gunicorn==21.2.0
python-dotenv==1.0.0
